implement try-either for clojurescript
try-either was previously not implemented for clojurescript, causing
try-success-test and try-exception-test to fail.

This requires the try-either macro to detect when it is being expanded
in a cljs namespace, as reader conditionals have already been expanded
by the time macros are evaluated.

See https://groups.google.com/forum/#!topic/clojurescript/iBY5HaQda4A
for a discussion of the approach used.