Type hints for core namespaces (#152)
Added type hints to avoid reflecting on fields at runtime. This should
provide a significant performance improvement for hot code.

Note, there is a trade-off with this approach: third party types will
no longer be useable in place of cats types because the field references
are no longer reflectively looked up by name but statically determined.
If this becomes a problem a better approach would be to add protocols for
retrieving shared values instead of direct field access.